Add `Result<V, E>` enum
- Add native `Result<V, E>` enum type in `prelude`!
- Refactor abra_native module to support creating enum types using the
existing proc macros. It's still in its pretty early stages, but all the
main pieces are in place.
- Add `Type::flatten` function to handle cases like
```
val results = [Result.Ok("asdf"), Result.Err(123)]
```
Previously, the type of `results` would be
`(Result<String, _> | Result<_, Int>)[]`, but now it gets flattened to a
much more reasonable representation: `Result<String, Int>[]`. This will
_defintely_ need to be iterated on, as it only works for `Placeholder`
types in `Reference` types at the moment.
